Delete unnecessary files; start test_client 1 second into sim
