+++ test/SIM_test_varserv/RUN_test/unit_test.py
@@ -11,14 +11,14 @@ def main():
 	trick.exec_set_software_frame(0.01)
 	# trick.set_var_server_info_msg_on()
 
-	trick.exec_set_terminate_time(5.0)
+	trick.exec_set_terminate_time(10.0)
 
 	varServerPort = trick.var_server_get_port()
 	test_output = ( os.getenv("TRICK_HOME") + "/trick_test/SIM_test_varserv.xml" ) 
+	command = 'os.system("./models/test_client/test_client ' + str(varServerPort) + ' --gtest_output=xml:' + test_output + ' &")'
 
-	# Start the test client
-	os.system("./models/test_client/test_client " + str(varServerPort) + " --gtest_output=xml:" + test_output + " &")
-
+	# Start the test client after everything has been initialized (hopefully)
+	trick.add_read(1.0, command)
 
 if __name__ == "__main__":
 	main()
