+++ test/SIM_test_varserv/models/varserv/include/VS.hh
@@ -2,7 +2,7 @@
 PURPOSE:                     ( Variable server test )
 REFERENCES:                  ( None )
 ASSUMPTIONS AND LIMITATIONS: ( None )
-PROGRAMMERS:                 ( (Lindsay Landry) (L3) (9-12-2013) )
+PROGRAMMERS:                 ( (Lindsay Landry) (L3) (9-12-2013) ) ( (Jackie Deans) (CACI) (11-30-2022) )
 ********************************************************************************/
 #include <wchar.h>
 #include <string.h>
@@ -45,7 +45,7 @@ class VSTest {
 		int success();
 		int fail();
 
-		char *status_messages[3] = {
+		const char *status_messages[3] = {
 			"Variable Server Test Success",
 			"Variable Server Test Failure",
 			"Client failed to connect"
