--- test/SIM_test_varserv/models/varserv/src/VS_init.cpp
@@ -1,18 +1,0 @@
-/******************************TRICK HEADER*************************************
-PURPOSE:                     ( Test variable server )
-REFERENCE:                   ( None )
-ASSUMPTIONS AND LIMITATIONS: ( None )
-CLASS:                       ( scheduled )
-LIBRARY DEPENDENCY:          ( VS.o )
-PROGRAMMERS:                 ( (Lindsay Landry) (L3) (9-12-2013) )
-*******************************************************************************/
-#include <iostream>
-#include "../include/VS.hh"
-#include "sim_services/VariableServer/include/variable_server_proto.h"
-#include "sim_services/VariableServer/include/VariableServer.hh"
-#include "sim_services/UnitTest/include/trick_tests.h"
-
-int VSTest::init() {
-	status = 2; 
-	return (0);
-}
